// app/shared/models/GrievanceStatus.model.ts
export class GrievanceStatusModel {
     nGrievanceStatusId: number = 0;
     tGrievanceStatus: string = null;
     nGrievanceStatusOrder: number = 0;
     bIsActive: boolean = false;
}

export class GrievanceStatusViewModel 
{
     nGrievanceStatusId: number = 0;
     tGrievanceStatus: string="";
     nGrievanceStatusOrder: number = 0;
     dCreatedDt : string = ""
     bIsActive: boolean = false;
}